requirements.txt
discord.py==1.6.0
youtube-dl==2021.2.10
async-timeout==3.0.1
asyncio==3.4.3
timeout==0.1.2
partial==1.0
discord.py[voice]
PyNaCl==1.4.0
ffmpeg==1.4
opuslib==3.0.1
discord.py[voice]==0.16.12